refactor: file jsToEsm.cjs and rename to esm.mjs
